requirements.txt
alembic==1.11.1
anyio==3.7.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
pyTelegramBotAPI==4.12.0
python-telegram-bot==20.3
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.18
telebot==0.0.5
typing_extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
